extract logic for exclusions
